Removed old unneeded ner parsing files
