+++ src/llama_cpp_chat_memory/document_parsing/spacier/__init__.py
@@ -1,1 +1,1 @@
-from document_parsing.spacier import core, extensions, utils
+from document_parsing.spacier import core, utils
